Allow specifying device (by serial number or other qualifier

+++ lib/chiizu/runner.rb
@@ -33,13 +33,37 @@ module Chiizu
         return
       end
 
+      devices = executor.execute(command: "adb devices -l", print_all: true, print_command: true).split("\n")
+      # the first output by adb devices is "List of devices attached" so remove that
+      devices = devices.drop(1)
+
+      if devices.empty?
+        UI.error 'There are no connected devices or emulators'
+        return
+      end
+
+      devices.select! { |d| d.include?(config[:specific_device]) } if config[:specific_device]
+      if devices.empty?
+        UI.error "No connected devices matched your criteria: #{config[:specific_device]}"
+        return
+      end
+
+      if devices.length > 1
+        UI.important "Multiple connected devices, selecting the first one"
+        UI.important "To specify which connected device to use, use the -s (specific_device) config option"
+      end
+
+      # grab the serial number. the line looks like this:
+      # 00c22d4d84aec525       device usb:2148663295X product:bullhead model:Nexus_5X device:bullhead
+      device_serial = devices[0].match(/^[\w]+/)[0]
+
       if config[:clear_previous_screenshots]
         UI.message "Clearing output directory of screenshots at #{config[:output_directory]}"
         files = Dir.glob(File.join('.', config[:output_directory], '**', '*.png'))
         File.delete(*files)
       end
 
-      device_ext_storage = executor.execute(command: "adb shell echo \\$EXTERNAL_STORAGE",
+      device_ext_storage = executor.execute(command: "adb -s #{device_serial} shell echo \\$EXTERNAL_STORAGE",
                                           print_all: true,
                                       print_command: true)
       device_screenshots_path = File.join(device_ext_storage, config[:app_package_name])
@@ -47,7 +71,7 @@ module Chiizu
       # TODO need to handle commands failing
 
       UI.message 'Cleaning screenshots directory on device'
-      executor.execute(command: "adb shell rm -rf #{device_screenshots_path}",
+      executor.execute(command: "adb -s #{device_serial} shell rm -rf #{device_screenshots_path}",
                      print_all: true,
                  print_command: true)
 
@@ -57,7 +81,7 @@ module Chiizu
       end
 
       UI.message 'Installing app APK'
-      executor.execute(command: "adb install -r #{app_apk_path}",
+      executor.execute(command: "adb -s #{device_serial} install -r #{app_apk_path}",
                      print_all: true,
                  print_command: true)
 
@@ -67,25 +91,25 @@ module Chiizu
       end
 
       UI.message 'Installing tests APK'
-      executor.execute(command: "adb install -r #{tests_apk_path}",
+      executor.execute(command: "adb -s #{device_serial} install -r #{tests_apk_path}",
                      print_all: true,
                  print_command: true)
 
       UI.message 'Granting the permission necessary to change locales on the device'
-      executor.execute(command: "adb shell pm grant #{config[:app_package_name]} android.permission.CHANGE_CONFIGURATION",
+      executor.execute(command: "adb -s #{device_serial} shell pm grant #{config[:app_package_name]} android.permission.CHANGE_CONFIGURATION",
                      print_all: true,
                  print_command: true)
 
-      device_api_version = executor.execute(command: "adb shell getprop ro.build.version.sdk",
+      device_api_version = executor.execute(command: "adb -s #{device_serial} shell getprop ro.build.version.sdk",
                                           print_all: true,
                                       print_command: true).to_i
 
       if (device_api_version >= 23)
         UI.message 'Granting the permissions necessary to access device external storage'
-        executor.execute(command: "adb shell pm grant #{config[:app_package_name]} android.permission.WRITE_EXTERNAL_STORAGE",
+        executor.execute(command: "adb -s #{device_serial} shell pm grant #{config[:app_package_name]} android.permission.WRITE_EXTERNAL_STORAGE",
                        print_all: true,
                    print_command: true)
-        executor.execute(command: "adb shell pm grant #{config[:app_package_name]} android.permission.READ_EXTERNAL_STORAGE",
+        executor.execute(command: "adb -s #{device_serial} shell pm grant #{config[:app_package_name]} android.permission.READ_EXTERNAL_STORAGE",
                        print_all: true,
                    print_command: true)
       end
@@ -93,7 +117,7 @@ module Chiizu
       config[:locales].each do |locale|
         UI.message "Running tests for locale: #{locale}"
 
-        instrument_command = ["adb shell am instrument --no-window-animation -w",
+        instrument_command = ["adb -s #{device_serial} shell am instrument --no-window-animation -w",
                               "-e testLocale #{locale.gsub("-", "_")}",
                               "-e endingLocale #{config[:ending_locale].gsub("-", "_")}"]
         instrument_command << "-e class #{test_classes_to_use.join(',')}" if test_classes_to_use
@@ -106,7 +130,7 @@ module Chiizu
       end
 
       UI.message "Pulling captured screenshots from the device"
-      executor.execute(command: "adb pull #{device_screenshots_path}/app_lens #{config[:output_directory]}",
+      executor.execute(command: "adb -s #{device_serial} pull #{device_screenshots_path}/app_lens #{config[:output_directory]}",
                      print_all: true,
                  print_command: true)
